Coverity ID #503928
Commit 99ed12d949 added a speculative overflow detection for extra
protection, but I forgot to cast the type in the calculation which
leads to truncation of the value.

Add a cast to the larger data type to prevent that.

+++ base/gsicc_create.c
@@ -1448,7 +1448,7 @@ gsicc_create_initialize_clut(gsicc_clut *clut)
     clut->clut_num_entries = entries = clut->clut_dims[0];
 
     for (k = 1; k < clut->clut_num_input; k++) {
-        entries = clut->clut_num_entries * clut->clut_dims[k];
+        entries = (int64_t)clut->clut_num_entries * clut->clut_dims[k];
         if (entries > INT_MAX || entries / clut->clut_num_entries != clut->clut_dims[k]) {
             clut->clut_num_entries = 0;
             break;
